Refactor class organization for enums, models, and type aliases
