+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/model/Difficulty.kt
@@ -1,0 +1,8 @@
+package dev.teogor.sudoklify.model
+
+enum class Difficulty {
+  EASY,
+  MEDIUM,
+  HARD,
+  EXPERT,
+}
