+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/model/Sudoku.kt
@@ -1,0 +1,12 @@
+package dev.teogor.sudoklify.model
+
+import dev.teogor.sudoklify.types.PuzzleString
+import dev.teogor.sudoklify.types.SolutionString
+
+
+data class Sudoku(
+  val puzzle: PuzzleString,
+  val solution: SolutionString,
+  val difficulty: Difficulty,
+  val type: Type,
+)
