+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/types/Board.kt
@@ -1,0 +1,3 @@
+package dev.teogor.sudoklify.types
+
+typealias Board = Array<Array<Cell>>
