+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/types/Cell.kt
@@ -1,0 +1,3 @@
+package dev.teogor.sudoklify.types
+
+typealias Cell = String
