+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/types/PuzzleString.kt
@@ -1,0 +1,3 @@
+package dev.teogor.sudoklify.types
+
+typealias PuzzleString = SudokuString
