+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/types/SudokuString.kt
@@ -1,0 +1,3 @@
+package dev.teogor.sudoklify.types
+
+typealias SudokuString = String
