+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/types/TokenMap.kt
@@ -1,0 +1,3 @@
+package dev.teogor.sudoklify.types
+
+typealias TokenMap = Map<Token, Cell>
